+++ worktile-server/worktile-logic/src/main/java/com/sjtu/worktile/model/TTaskAttachment.java
@@ -3,11 +3,11 @@ package com.sjtu.worktile.model;
 import java.util.Date;
 
 public class TTaskAttachment {
-    private Integer id;
+    private Long id;
 
-    private Integer taskId;
+    private Long taskId;
 
-    private Integer userId;
+    private Long userId;
 
     private String description;
 
@@ -27,27 +27,27 @@ public class TTaskAttachment {
 
     private String pending3;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getTaskId() {
+    public Long getTaskId() {
         return taskId;
     }
 
-    public void setTaskId(Integer taskId) {
+    public void setTaskId(Long taskId) {
         this.taskId = taskId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
